Changed Class MachineServiceImpl. Removed comment. The class implements the MachineService interface, typed <Machine>. The class-level ProducerStrategy object has been added and initialized.
Overridden method List<Machine> getAll(Class<? extends Machine> type).
The body of the method is wrapped in a try-catch block. Inside, we declare a producer object, which we initialize with the appropriate type using a getter, accessing the contents of the HashMap and calling the get method, passing inside the type.
After that, we declare a List and populate it with the appropriate machine objects using producer.get().
We return a new ArrayList, passing it to the generated List.
If an invalid type was passed to us, we will not be able to initialize the List and instead of returning a RuntimeException error will be thrown, and an empty ArrayList will be returned as a response.

Overridden method void fill(List<? super Machine> machines, Machine value).
Inside, a loop is declared that goes through the entire List and replaces each element with a value.

Overridden method void startWorking(List<? extends Machine> machines).
Inside, a loop is declared that goes through the entire List of machines and calls the doWork () method for each of them.