Added Picker interface. Internally, we initialized the Random object;
Added a static method `random` typed by the producer `<E extends Picker>` so that we do not create an array of `objects`.
The random() method takes all the ENUM values as an array E and then returns the element from the ENUM as a string using a randomly generated index.

The String name() method is the standard ENUM method for returning a value as a string. This decision is also justified by the fact that the use of toString() is not welcome in business logic.