#2
class Excavator has two private fields.The class has two constructors. The first one is a parameterless constructor that calls the constructor of the parent class, Machine. The second one calls the constructor of the parent class with parameters - name and color - and also initializes the fields of bucketCapacity,tracksOrWheels corresponding passed values.