#3
I create a couple of variables that are specific to the Truck class. The first constructor is parameterless and calls the constructor from the parent class, Machine. The second constructor calls constructor from the parent class with the parameters of name and color, and initializes the fields that belong to the Truck class.