#4
Now this method, instead of returning a regular list, returns a list of type "List<? extends T>". This means that any type T that will be used in implementing this interface must be a subtype of the Machine class. And the get() method now returns a list of objects that inherit from T. Using a wildcard in the type definition of the list allows to return a list that contains any types that are subtypes of T, not just objects of type T.