#7
"I created 3 final variables with the necessary values that will be used later in switch statment.

In the getAll() method, i initially wanted to use an if-else statement, but decided that a switch would be easier to understand. This code determines which type of machine is passed to getAll() using the type argument, and then an object of the corresponding class is created. If the passed type does not match, an empty list is returned.

the fill method populates the machines list with the value, using an upper bounded wildcard <? super Machine>, which allows objects of type Machine and its superclasses to be added to the list.

 And finally, the startWorking() method starts the operation for each machine in the list.